Refactored code further to allow easier DI wire up of IKeysProvider implementations via ServiceExtension methods.
