+++ src/AKSoftware.Localization.MultiLanguages/ServiceExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using System;
+using Microsoft.Extensions.DependencyInjection;
 using System.Globalization;
 using System.Reflection;
 
@@ -40,27 +41,25 @@ namespace AKSoftware.Localization.MultiLanguages
             });
         }
 
-
         /// <summary>
-        /// Register a singleton instance of LanguageContainer class initialized from the executing assembly
+        /// Register a singleton instance of LanguageContainer class initialized with the user culture
         /// </summary>
         /// <param name="services">Dependency Services provider</param>
+        /// <param name="assembly">Assembly that contains the Resource folder which has the language files</param>
         /// <returns></returns>
-        //public static IServiceCollection AddLangaugeContainer(this IServiceCollection services)
-        //{
-        //    return services.AddSingleton<ILanguageContainerService, LanguageContainerInAssembly>(s => new LanguageContainerInAssembly(Assembly.GetEntryAssembly()));
-        //}
+        public static IServiceCollection AddLanguageContainer<TKeysProvider>(this IServiceCollection services, Assembly assembly, string folderName = "Resources")
+        where TKeysProvider : KeysProvider
+        {
+            services.AddSingleton<IKeysProvider, TKeysProvider>(s => (TKeysProvider)Activator.CreateInstance(typeof(TKeysProvider), assembly, folderName));
+            return services.AddSingleton<ILanguageContainerService, LanguageContainerInAssembly>(s =>
+            {
+                var keysProvider = s.GetService<IKeysProvider>();
+                return new LanguageContainerInAssembly(keysProvider);
+            });
+        }
 
-        /// <summary>
-        /// Register a singleton instance of LanguageContainer class initialized from the executing assembly
-        /// </summary>
-        /// <param name="services">Dependency Services provider</param>
-        /// <param name="defaultCulture">Default Culture</param>
-        /// <returns></returns>
-        //public static IServiceCollection AddLangaugeContainer(this IServiceCollection services, CultureInfo defaultCulture)
-        //{
-        //    return services.AddSingleton<ILanguageContainerService, LanguageContainerInAssembly>(s => new LanguageContainerInAssembly(Assembly.GetEntryAssembly(), defaultCulture));
-        //}
+
+       
 
     }
 }
